PhishingKitAnalyzer/requirements.txt
yara-python>=4.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.65.0
python-magic>=0.4.27
aiohttp>=3.8.5
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0.1
rich>=13.7.0
docker>=7.0.0
pdfkit>=1.0.0 
